+++ src/parser/machine/stack.rs
@@ -1,0 +1,8 @@
+pub struct Stack;
+
+impl Stack {
+    pub fn new() -> Self {
+        Stack
+    }
+}
+
