+++ src/parser/mod.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod syntax;
 mod machine;
